completed playSong function and added time converter function.
I completed the play song function by first isolating the song by id using the filter method. then I called the play song function in the player object, and I inserted the isolated song as an argument. Then I made the player.playSong function print the song that was playing, in the correct format. I created a time converter function to print the time in the correct format.