removeSongFunction
first I changed the names of some things to be more descriptive like timeCoverter to convertToMin.

Then I did the remove song function.

first, I made it throw an error if the id of the song trying to be removed doesn't  exists.

second, then I filtered the song out of the songs array.

third I filtered the song out of each playlist.