addSongFunction
I completed the addSong Function.

first, I threw an error if the id of the song trying to be added already exists.

second, I gave a default value to the id of the song being added and made sure that that id wasn't already taken.

third, I added the new song object to the array songs in player. I had to make another function called convertToSec, to get the duration to the correct format.

fourth, I returned the id like the instructions said to do.