removePlaylist function.
I changed the name of obj variable, within a filter method, to song, in order to make it easier to understand.

Then I completed the remove function.

first, I made sure the function would throw an error if the id of the function being removed didn't exist.

second, I filtered the playlist out of the out of the playlists key