remove song
first, i used the findID function in order to throw an error if the id doesn't exist. second, i removed the song with the specific ID from our sons array. i created a for loop that would run on each cell inside the songs array. then i asked if the id of the specific cell is the same as the id that function recived. if so, we will use splice to remove the cell from the array. at last, the function also needs to to remove the song from the playlist. in order to do that we use 2 for loops. the first loop will run on each playlist and the second loop will run on the song array that we have in each playlist. inside the second for loop we need to see if the id that the function recived is equal to one of the ID's in the array song. if so, we will use splice to remove it.