add song
here i wrote the function addsong. the function should throw if an ID is already exist. also the function supposed to add a song to song array in case the ID is not taken. so in the function I created a new song object that his values is the values that the function recived from the user. also the function recive duartion in the fomat mm:ss and expected to cnvert it to numbers. so I created the function convertSeconds to resolve that. also we need to give a default ID in case the function didn't recive so here i gave ID=15.