delete playlist
this function needs to delete a playlist , it's spouse to check if the ID exists and if not throw a statement. also it should delete the the playlist with the same ID that the function recived. i created a function(findIDplaylist) that checks the ID's inside the playlists key and if there is a match the function would return true. so by that logic, we want to see if the function will give us false and that's sign that there is no match between the playlists ID's and the ID that the function recived and we will throw a statement. if the ID exist we will use a for loop to run on key playlists and check each playlist inside playlists and we will see if the playlist id match's to the ID that the function recived. if so, we would delete the specific playlist and we would subtract the size of playlists.