add playlist
the function suppose to add a playlist to player. first, we need to check if the playlist ID is not taken. so we will use the findIdPlaylist to see if the ID that function recived is equal to one of the ID's that playlists has. if the function return true then we will throw a statement the you cant use this ID. secondly, we need to add a new playlist to playlists. so the function will create a new playlist inside her and we will use push to add it to playlists.