playPlaylist
the function needs to do 2 things. first, the function checks if the ID that it got is even exist inside playlists. we use the function findIdPlaylist to see if the id exists if no it will throw an error. second, we need to see inside playlist which playlist has the same ID to the one the function recived. so we will create a for loop that will run on each playlist and check which one of them has the same ID as the one that function. when we find that specific playlist we create another for loop that will run on the playlist songs. and we use the function playSong to play each song.