edit playlist
this function needs to do 5 thing. the first one is to check if the  playlist ID that it got from the user exsits. we can check it with findIdPlaylist function. the scond thing is to check if a song ID is inside the songs array. here we can use the find ID function to help us. the third , fourth and fifth thing that the function has to do is to check if a specific playlist(playlist with the id that the function recived) has a specific song(song with the id that the function recvied) and to do the following things: if the song is inside the playlist, remove it. it would be done by the second if condition in the for loops. if the songs key inside playlist has one cell and it's value is the songID. we have to delete the entire playlist. it would be done by the first if statement in the for loops. and the last thing we need to do is to push the song in to the playlist if it isn't there, which would be done by the last if statement inside the for loops.