sum duration
this function receives an ID of a playlist and needs to return the duration of all the songs in it. so, at first, we need to find the playlist with the same ID as the one that the function recvied. so to that we create a for loop that will run on playlist and find the one we are looking for. after we found the right playlist we need to make another for loop that would run on the playlist songs. in this for loop we would have another for loop. the last for loop would run on songs and will find us the song with the same ID as the one in the songs in playlists. after we found match we will push to sum the duration of the specific song. the last part of the program(the one when we compare ID'S of playlists songs and songs) will run until we gone over all the songs in playlist.