search by query
the function receive a string. it needs to return a new object player that has all the songs and playlists that include this string and the check must be insensitive. so at first we make a for loop that checks if the string is inside the array songs, if it is in the array then we will push the specific song in to our returned player. we are also doing the same thing inside the playlists array. after that we have everything inside our returned player but we need to organize him alphanumrly before we return him. so, in order to that we will make 2 arrow function(one runs on songs and the other on playlists), each one with compare function inside. this will put the objects inside playlists and songs in the right order. after that we return our new player.