playPlaylist Function
I simultaneously checked to make sure that the id does exist and provided the index of that id in playlists array. Then I used the forEach method on the array of song ids in the playlist to put them all through the playSong function. This plays all the songs in that playlist.