edit Playlist Function
First, I check that the songId does exist. Then, I simultaneously get the exact playlist I'm working with and check the playlistId.

Second, I say if the songId is already in the playlist to filter it out of the playlist.

Third, the else statement states that if the songId is not in the playlist to add it.

fourth, I remove the playlist if it is empty as a result of this function.