search by duration
this function receive a  specific duration and needs to return the song or the playlist with the closest duration. I created an array that each cell of his is an object that has 3 keys: 1) the duration of a specific song.2) the range between this song duration and the duration that function recived.3) this specific song so we could return him with quick acsess. I also created the same array that would work on playlists. then, I used the sort operator  to arrange the arrays so that the object which is closest to duration that the function recived would be at index 0. and after that we just compare the ranges between the playlist and the song(that are located at index 0) , the one with the lowest range is the closest to duration and we would return him.