better editing + function accesibility (customizing)
edited it nicer and logged to the console all the functions in the project. (ones i made and ones which came with the project), thus it be easier for future programmers to see all functions and what they do in the console (when no error occurs) without needing to scroll and search for specific function in this big code. :D ty